pyproject.toml
[tool.poetry]
name = "python-quiz"
version = "0.1.0"
description = ""
authors = ["Michael <mpgg2000@gmail.com>"]
readme = "README.md"
packages = [{include = "python_quiz"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^1.10.5"
pytelegrambotapi = "^4.10.0"
aiohttp = "^3.8.4"
yoyo-migrations = "^8.2.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.8"}
mypy = "^1.0.1"
python-dotenv = "^0.21.1"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
python-multipart = "^0.0.5"
requests = "^2.28.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
